wayzata1976_app/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
better-exceptions==0.3.2
black==20.8b1
boto3==1.17.0
botocore==1.20.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
Collectfast==2.2.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.5
django-bmemcached==0.3.0
django-debug-toolbar==3.2
django-heroku==0.3.1
django-material-icons==0.1.2
django-s3direct==1.1.8
django-s3file==5.2.0
django-storages==1.11.1
django-tinymce==3.2.0
django-upload-form==0.4.3
django-widget-tweaks==1.4.8
fontawesome-free==5.15.2
geographiclib==1.50
geopy==2.1.0
google-api-core==1.26.0
google-api-python-client==2.0.2
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.19.0
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
msgpack==0.6.2
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.3
pathspec==0.8.1
pep517==0.8.2
Pillow==8.1.0
pip-search==0.0.2
progress==1.5
protobuf==3.15.3
psycopg2==2.8.6
# psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-binary-memcached==0.30.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.5
redis==3.5.3
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
s3transfer==0.3.4
six==1.14.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tabulate==0.8.7
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uhashring==2.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
